requirements.txt
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
configobj==5.0.6
constantly==15.1.0
distro==1.4.0
entrypoints==0.3
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.24.3
oauthlib==3.1.0
pandas==2.0.1
pbr==5.11.1
pexpect==4.6.0
pulsar-client==3.1.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.23.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
pyOpenSSL==23.1.1
pyrsistent==0.15.5
pyserial==3.4
pymongo==4.3.3
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
ssh-import-id==5.10
supervisor==4.1.0
systemd-python==234
testresources==2.0.1
tqdm==4.65.0
Twisted==18.9.0
tzdata==2023.3
